Add exact number of found solutions stop criterion and change type of number of solutions from u16 to usize

+++ oxigen/src/lib.rs
@@ -237,12 +237,10 @@ impl<T, Ind: Genotype<T>> GeneticExecution<T, Ind> {
             self.print_progress_header();
         }
 
-        while !self.stop_criterion.stop(
-            generation,
-            progress,
-            solutions.len() as u16,
-            &current_fitnesses,
-        ) {
+        while !self
+            .stop_criterion
+            .stop(generation, progress, solutions.len(), &current_fitnesses)
+        {
             generation += 1;
 
             mutation_rate = self.mutation_rate.rate(
